+++ mitra_validators/src/posts.rs
@@ -8,7 +8,10 @@ use mitra_models::posts::types::{
 };
 use mitra_utils::html::{clean_html_strict, clean_html_all};
 
-use super::errors::ValidationError;
+use super::{
+    activitypub::validate_object_id,
+    errors::ValidationError,
+};
 
 pub const ATTACHMENT_LIMIT: usize = 15;
 pub const MENTION_LIMIT: usize = 50;
@@ -16,7 +19,6 @@ pub const HASHTAG_LIMIT: usize = 100;
 pub const LINK_LIMIT: usize = 10;
 pub const EMOJI_LIMIT: usize = 50;
 
-const OBJECT_ID_SIZE_MAX: usize = 2000;
 const TITLE_LENGTH_MAX: usize = 300;
 pub const CONTENT_MAX_SIZE: usize = 100000;
 const CONTENT_ALLOWED_TAGS: [&str; 8] = [
@@ -89,9 +91,7 @@ pub fn validate_post_create_data(
         return Err(ValidationError("too many emojis"));
     };
     if let Some(ref object_id) = post_data.object_id {
-        if object_id.len() > OBJECT_ID_SIZE_MAX {
-            return Err(ValidationError("object ID is too long"));
-        };
+        validate_object_id(object_id)?;
     };
     Ok(())
 }